+++ src/components/ENSListPage/ENSListPage.container.ts
@@ -2,14 +2,14 @@ import { connect } from 'react-redux'
 import { push } from 'connected-react-router'
 import { locations } from 'routing/locations'
 import { RootState } from 'modules/common/types'
-import { MapStateProps, MapDispatchProps, MapDispatch } from './NamesPage.types'
+import { MapStateProps, MapDispatchProps, MapDispatch } from './ENSListPage.types'
 import { getENSByWallet, getLoading } from 'modules/ens/selectors'
 import { isLoggingIn, isLoggedIn } from 'modules/identity/selectors'
 import { FETCH_ENS_LIST_SUCCESS } from 'modules/ens/actions'
 import { getAddress } from 'decentraland-dapps/dist/modules/wallet/selectors'
 import { getSortBy, getTotalPages, getPage } from 'modules/ui/ens/selectors'
 
-import NamesPage from './NamesPage'
+import ENSListPage from './ENSListPage'
 import { isLoadingType } from 'decentraland-dapps/dist/modules/loading/selectors'
 import { getLands } from 'modules/land/selectors'
 
@@ -26,7 +26,7 @@ const mapState = (state: RootState): MapStateProps => ({
 
 const mapDispatch = (dispatch: MapDispatch): MapDispatchProps => ({
   onNavigate: path => dispatch(push(path)),
-  onPageChange: options => dispatch(push(locations.names(options)))
+  onPageChange: options => dispatch(push(locations.enslist(options)))
 })
 
-export default connect(mapState, mapDispatch)(NamesPage)
+export default connect(mapState, mapDispatch)(ENSListPage)
