+++ src/components/ENSListPage/index.ts
@@ -1,0 +1,2 @@
+import ENSListPage from './ENSListPage.container'
+export default ENSListPage
