--- src/components/NamesPage/index.ts
@@ -1,2 +1,0 @@
-import NamesPage from './NamesPage.container'
-export default NamesPage
