+++ src/routing/locations.ts
@@ -27,5 +27,5 @@ export const locations = {
   collectionDetail: (collectionId = ':collectionId') => `/collections/${collectionId}`,
   itemEditor: (options: { itemId?: string; collectionId?: string } = {}) =>
     injectParams(`/item-editor`, { itemId: 'item', collectionId: 'collection' }, options),
-  names: (options: PaginationOptions = {}) => injectPagination('/names', options)
+  enslist: (options: PaginationOptions = {}) => injectPagination('/enslist', options)
 }
